Create skeleton of the action
